+++ client/repo.go
@@ -10,7 +10,7 @@ import (
 
 type ClientRepo interface {
 	// Define methods for interacting with the API
-	GetByID(id int) (client Client, err error)
+	GetByID(id int64) (client Client, err error)
 }
 type clientRepo struct {
 	db *sqlx.DB
@@ -20,7 +20,7 @@ func NewClientRepo(db *sqlx.DB) ClientRepo {
 	return &clientRepo{db: db}
 }
 
-func (r *clientRepo) GetByID(id int) (client Client, err error) {
+func (r *clientRepo) GetByID(id int64) (client Client, err error) {
 
 	err = r.db.Get(&client, `
 		select
