+++ client/service.go
@@ -8,7 +8,7 @@ func NewClientService(client ClientRepo) *ClientService {
 	return &ClientService{client: client}
 }
 
-func (s ClientService) GetProfile(id int) (client Client, err error) {
+func (s ClientService) GetProfile(id int64) (client Client, err error) {
 	client, err = s.client.GetByID(id)
 	if err != nil {
 		return client, err
